+++ hooks/src/lib/index.ts
@@ -1,5 +1,5 @@
-export { useCardPasswordInput } from './useCardPassword';
-export { useCardCVCInput } from './useCardCVC';
-export { useCardExpDateInput } from './useCardExpDate';
-export { useCardNumbersInput } from './useCardNumbers';
-export { useCardCompanyInput } from './useCardCompany';
+export { useCardPasswordInput } from './useCardPasswordInput';
+export { useCardCVCInput } from './useCardCVCInput';
+export { useCardExpDateInput } from './useCardExpDateInput';
+export { useCardNumbersInput } from './useCardNumbersInput';
+export { useCardCompanyInput } from './useCardCompanyInput';
