+++ hooks/src/lib/useCardCVCInput.test.ts
@@ -1,5 +1,5 @@
 import { renderHook, act } from '@testing-library/react';
-import { useCardCVCInput } from './useCardCVC';
+import { useCardCVCInput } from './useCardCVCInput';
 import { ChangeEvent } from 'react';
 import { ERROR_MESSAGE } from './validator/constants/errorMessage';
